requirements.txt
Django==2.2.1
django-autocomplete-light==3.3.4
django-nested-admin==3.2.3
python-monkey-business==1.0.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
